refactor: correct imports and property names
