+++ packages/deployment-service/src/utils/index.ts
@@ -1,3 +1,3 @@
-export * from './authorised'
 export * from './ownership'
 export * from './resolve-customer'
+export * from './resolve-developer-id'
